pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "app"
version = "0.1.0"
description = "DDD-based FastAPI project skeleton"
requires-python = ">=3.11"
authors = [
  { name = "Developer" }
]
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0",
  "sqlalchemy>=2.0.25",
  "alembic>=1.13.0",
  "pydantic-settings>=2.1.0",
  "psycopg[binary,pool]>=3.1.18",
  "structlog>=24.1.0",
  "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.26.0",
  "faker>=22.0.0",
  "ruff>=0.1.0",
  "mypy>=1.8.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
